Great Job
Hi Weihong and Yuan,

Really clean and usable application, the code is nicely done and the modules and folder structure and organization is very good. Good job on the CRUD functionalities of this project, it is well done and it works nicely. I left some minor comments on the HTML tags, some tips on the presentation style of the repo's readme (using gifs), and some other minor questions/comments. 

Very good job!
Elliot